backend/requirements.txt
# --- Web framework ---
fastapi==0.121.1
uvicorn[standard]==0.38.0
starlette==0.49.3
httptools==0.7.1
h11==0.16.0
watchfiles==1.1.1

# --- Database ---
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11      # PostgreSQL driver
PyMySQL==1.1.2                # MySQL driver

# --- Environment and config ---
python-dotenv==1.2.1

# --- Web scraping ---
requests==2.32.5
beautifulsoup4==4.14.2
lxml==4.9.3
soupsieve==2.8
charset-normalizer==3.4.4
urllib3==2.5.0
certifi==2025.10.5
idna==3.11

# --- Data validation and models ---
pydantic==2.12.4
typing-extensions==4.15.0
orjson==3.11.4

# --- LangChain and LLMs ---
langchain-core==1.0.4
langchain-community==0.4.1
langchain-google-genai==3.0.1
langchain-classic==1.0.0
langchain-text-splitters==1.0.0
langsmith==0.4.41
google-ai-generativelanguage==0.9.0
httpx==0.28.1
httpcore==1.0.9
aiohttp==3.13.2
anyio==4.11.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
sniffio==1.3.1
yarl==1.22.0
multidict==6.7.0

# --- Utilities ---
python-multipart==0.0.6

# --- Production-ready server ---
gunicorn==21.2.0
uvloop==0.22.1


cachetools==5.6.2










# fastapi
# uvicorn[standard]
# SQLAlchemy
# psycopg2-binary
# pymysql
# python-dotenv
# requests
# beautifulsoup4
# langchain-core
# langchain-community
# langchain-google-genai
# pydantic